requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.3
black==25.11.0
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
flake8==7.3.0
graphql-core==3.2.7
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
kombu==5.5.4
lia-web==0.2.3
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
strawberry-graphql==0.287.0
strawberry-graphql-django==0.67.1
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
